requirements.txt
chris_plugin==0.4.0
pillow
pdf2image

